refactor md location
